Change UI class/fxml name from person to contact
